Revert "add Internationalization"
