+++ app/components/dynamicBlur.tsx
