+++ app/layout.tsx
@@ -1,0 +1,59 @@
+import type { Metadata } from "next";
+import { Inter } from "next/font/google";
+import Navbar from "./components/navbar";
+import "./globals.css";
+const inter = Inter({ subsets: ["latin"] });
+
+const APP_NAME = "mangaboiboi";
+const APP_DEFAULT_TITLE = "MangaBoiBoi";
+const APP_TITLE_TEMPLATE = "%s | MangaBoiBoi";
+const APP_DESCRIPTION = "mangaBoiBoi";
+export const metadata: Metadata = {
+  applicationName: APP_NAME,
+  title: {
+    default: APP_DEFAULT_TITLE,
+    template: APP_TITLE_TEMPLATE,
+  },
+  description: APP_DESCRIPTION,
+  appleWebApp: {
+    capable: true,
+    statusBarStyle: "default",
+    title: APP_DEFAULT_TITLE,
+    // startUpImage: [],
+  },
+  formatDetection: {
+    telephone: false,
+  },
+  openGraph: {
+    type: "website",
+    siteName: APP_NAME,
+    title: {
+      default: APP_DEFAULT_TITLE,
+      template: APP_TITLE_TEMPLATE,
+    },
+    description: APP_DESCRIPTION,
+  },
+  twitter: {
+    card: "summary",
+    title: {
+      default: APP_DEFAULT_TITLE,
+      template: APP_TITLE_TEMPLATE,
+    },
+    description: APP_DESCRIPTION,
+  },
+};
+
+export default function RootLayout({
+  children,
+}: Readonly<{
+  children: React.ReactNode;
+}>) {
+  return (
+    <html lang="en">
+      <body className={inter.className}>
+        <Navbar />
+        {children}
+      </body>
+    </html>
+  );
+}
