+++ app/scrollbar.css
@@ -1,0 +1,17 @@
+/* scrollbar.css */
+::-webkit-scrollbar {
+  height: 10px;
+}
+
+::-webkit-scrollbar-track {
+  background: transparent;
+}
+
+::-webkit-scrollbar-thumb {
+  background: rgba(0, 94, 255, 0.586);
+  border-radius: 20px;
+}
+
+::-webkit-scrollbar-thumb:hover {
+  background: rgba(26, 36, 232, 0.5);
+}
