--- i18n.json
@@ -1,8 +1,0 @@
-{
-  "locales": ["en", "fr"],
-  "defaultLocale": "fr",
-  "pages": {
-    "*": ["common"],
-    "/[lang]": ["home"]
-  }
-}
