--- src/app/components/LocaleSwitcher.tsx
@@ -1,18 +1,0 @@
-import { locales } from "@/src/app/navigation";
-import { useLocale, useTranslations } from "next-intl";
-import LocaleSwitcherSelect from "./LocaleSwitcherSelect";
-
-export default function LocaleSwitcher() {
-  const t = useTranslations("LocaleSwitcher");
-  const locale = useLocale();
-
-  return (
-    <LocaleSwitcherSelect defaultValue={locale} label="">
-      {locales.map((cur) => (
-        <option key={cur} value={cur}>
-          {cur}
-        </option>
-      ))}
-    </LocaleSwitcherSelect>
-  );
-}
