--- src/app/components/navigationLink.tsx
@@ -1,23 +1,0 @@
-"use client";
-
-import { useSelectedLayoutSegment } from "next/navigation";
-import { ComponentProps } from "react";
-import { Link } from "../navigation";
-
-export default function NavigationLink({
-  href,
-  ...rest
-}: ComponentProps<typeof Link>) {
-  const selectedLayoutSegment = useSelectedLayoutSegment();
-  const pathname = selectedLayoutSegment ? `/${selectedLayoutSegment}` : "/";
-  const isActive = pathname === href;
-
-  return (
-    <Link
-      aria-current={isActive ? "page" : undefined}
-      href={href}
-      style={{ fontWeight: isActive ? "bold" : "normal" }}
-      {...rest}
-    />
-  );
-}
