--- src/app/navigation.ts
@@ -1,7 +1,0 @@
-import { createSharedPathnamesNavigation } from "next-intl/navigation";
-
-export const locales = ["en", "fr"] as const;
-export const localePrefix = "always"; // Default
-
-export const { Link, redirect, usePathname, useRouter } =
-  createSharedPathnamesNavigation({ locales, localePrefix });
