--- src/app/noscrollbar.css
@@ -1,19 +1,0 @@
-::-webkit-scrollbar {
-  width: 0px;
-  background-color: transparent;
-}
-
-/* Track */
-::-webkit-scrollbar-track {
-  background-color: transparent;
-}
-
-/* Handle */
-::-webkit-scrollbar-thumb {
-  background-color: transparent;
-}
-
-/* Handle on hover */
-::-webkit-scrollbar-thumb:hover {
-  background-color: transparent;
-}
