--- src/i18n.ts
@@ -1,14 +1,0 @@
-import { getRequestConfig } from "next-intl/server";
-import { notFound } from "next/navigation";
-
-// Can be imported from a shared config
-const locales = ["en", "fr"];
-
-export default getRequestConfig(async ({ locale }) => {
-  // Validate that the incoming `locale` parameter is valid
-  if (!locales.includes(locale as any)) notFound();
-
-  return {
-    messages: (await import(`../messages/${locale}.json`)).default,
-  };
-});
